+++ tools/build.sh
@@ -1,2 +1,9 @@
+#!/bin/sh
+
+set -e
+
 cd ..
+
 Rscript tools/build.R
+
+cd -