+++ NEWS.md
@@ -1,5 +1,24 @@
 # CHANGELOG
 
+## 0.2.1
+
+* Support custom runner #9
+* support env vars for windows platform #11
+* support flag to print every command before invocation #12
+* markdown documentation hosted on github #10
+* Consistent error handling for all platforms
+
+## 0.2.0
+
+* execute_script function renamed to execute
+* Add support for command line arguments to be passed to the script #1
+* Merge all output text instead of vector #5
+* Add status 0 instead of null #6
+* Pass env variables as part of script execution #7
+* add wait support #8
+* Export all functions #4
+* Add demos #3
+
 ## 0.1.0
 
 * Initial Release